base-stats, python: move stat processing out of serializable stat
Python Stats processing is inherent to the type of stat-processing visitor
rather than the stat itself. Therefore we move stat-porcessing out of the
serializable stats class and into each visitor to determine how each
statistic type should be dealt with. This simplifies the process for
defining new output formats as we now just describe the required
processing for each case type individually:

Scalar, Distribution and SparseHist for specfic values.
Vector and Vector 2d to hold sets of result values.
SimObjectGroups as well as the overall SimStat object in order to
manage the overall heirarchical structure of statistics.

Finally we define a dump method to write the processed statistics to
a file.

Change-Id: I4210090fd860908f14702d0a049068a8c8e48e34
Reviewed-by: Giacomo Travaglini <giacomo.travaglini@arm.com>